auto_finance/requirements.txt
# Auto Finance 고도화 시스템 의존성

# 기본 라이브러리
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3
urllib3==2.0.7
fake-useragent==1.4.0

# AI 모델
google-generativeai==0.3.2
openai==1.3.7
anthropic==0.7.8
transformers==4.35.2
torch==2.1.1

# 데이터 처리
pandas>=1.3.0
numpy==1.24.3
scikit-learn==1.3.2
yfinance==0.2.28
ta==0.10.2

# 자연어 처리
nltk==3.8.1
textblob==0.17.1
spacy==3.7.2
jieba==0.42.1
konlpy==0.6.0
vaderSentiment==3.3.2

# 웹 프레임워크
fastapi==0.104.1
uvicorn==0.24.0
dash==2.14.2
dash-bootstrap-components==1.5.0

# 데이터베이스
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# 캐싱 및 작업 큐
celery==5.3.4
cachetools==5.3.2

# 비동기 처리
aiohttp==3.9.1
asyncio

# 시각화
matplotlib==3.8.2
seaborn==0.13.0
plotly>=5.0.0

# 유틸리티
python-dotenv==1.0.0
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3
holidays==0.30

# 로깅 및 모니터링
loguru==0.7.2
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk==1.38.0

# 테스트
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 개발 도구
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# 추가 패키지
jinja2==3.1.2
markdown==3.5.1
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
Pillow==10.1.0
html5lib==1.1
scipy==1.11.4
